+++ springProject/src/main/java/com/shareMemo/demo/service/MemoService.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface MemoService {
-    public List<Memo> getMemoList(Integer memberId, Integer notebookId);
-    public Memo addMemo(Integer memberId, Integer notebookId, MemoDto memoDto);
-    public Memo deleteMemo(Integer memberId, Integer memoId);
+    public List<Memo> getMemoList(Long memberId, Integer notebookId);
+    public Memo addMemo(Long memberId, Integer notebookId, MemoDto memoDto);
+    public Memo deleteMemo(Long memberId, Integer memoId);
 }
